Calculate DPS

+++ RPGCharacters/WeaponAttributes.cs
@@ -3,6 +3,6 @@
     public class WeaponAttributes
     {
         public int Damage { get; set; }
-        public int AttackSpeed { get; set; }
+        public double AttackSpeed { get; set; }
     }
 }
